Ajout de la validation des entrées pour les routes des personnages et des utilisateurs, création de la route d'inscription pour les utilisateurs.

+++ Backend/routes/characters.js
@@ -3,6 +3,69 @@ const router = express.Router();
 const db = require("../config/db");
 const { check, validationResult } = require("express-validator");
 
+// validations
+const validId = [
+  check("id")
+    .isLength({ min: 20, max: 20 })
+    .withMessage("L'ID doit avoir 20 charactères")
+    .matches(/^[a-zA-Z0-9]+$/)
+    .withMessage("L'ID ne peut contenir que des lettres et des chiffres"),
+];
+let validations = [
+  check("name").trim().notEmpty().withMessage("Le nom est requis."),
+  check("genre")
+    .trim()
+    .notEmpty()
+    .isLowercase()
+    .withMessage("Le genre est requis."),
+  check("class").trim().notEmpty().withMessage("La classe est requis."),
+  check("race").trim().notEmpty().withMessage("La race est requis."),
+  check("alignement")
+    .trim()
+    .notEmpty()
+    .withMessage("L'alignement' est requis."),
+  check("Traits.PersonalityTraits")
+    .trim()
+    .notEmpty()
+    .withMessage("Les Traits de personalités sont requis."),
+  check("Traits.Ideals")
+    .trim()
+    .notEmpty()
+    .withMessage("Les Ideaux sont requis."),
+  check("Traits.Bonds")
+    .trim()
+    .notEmpty()
+    .withMessage("Les Obligations sont requis."),
+  check("Traits.Flaws")
+    .trim()
+    .notEmpty()
+    .withMessage("Les Défauts sont requis."),
+  check("statistics.strength")
+    .trim()
+    .notEmpty()
+    .withMessage("La force est requise."),
+  check("statistics.dexterity")
+    .trim()
+    .notEmpty()
+    .withMessage("La dextérité est requise."),
+  check("statistics.constitution")
+    .trim()
+    .notEmpty()
+    .withMessage("La constitution est requise."),
+  check("statistics.intelligence")
+    .trim()
+    .notEmpty()
+    .withMessage("L'intelligence est requise."),
+  check("statistics.wisdom")
+    .trim()
+    .notEmpty()
+    .withMessage("La sagesse est requise."),
+  check("statistics.charisma")
+    .trim()
+    .notEmpty()
+    .withMessage("Le charisme est requis."),
+];
+
 /**
  * route GET pour trouver la liste de tous les personnages
  */
@@ -13,11 +76,11 @@ router.get(
     check("orderDirection").escape().trim().optional().isIn(["asc", "desc"]),
   ],
   async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
     try {
-      const result = validationResult(req);
-      if (!result.isEmpty()) {
-        return res.status(400).json({ msg: "données invalides" });
-      }
       let {
         limit = 10,
         start = 0,
@@ -126,7 +189,12 @@ statistics.forEach((statistic) => {
 /**
  * route pour trouver le personnage avec le id
  */
-router.get("/:id", async (req, res) => {
+
+router.get("/:id", validId, async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
   const { id } = req.params;
 
   try {
@@ -150,79 +218,26 @@ router.get("/:id", async (req, res) => {
 /**
  * route pour enrengistrer les informations du personnages dans la bdd.
  */
-router.post(
-  "/",
-  [
-    check("name").trim().notEmpty().withMessage("Le nom est requis."),
-    check("genre").trim().notEmpty().withMessage("Le genre est requis."),
-    check("class").trim().notEmpty().withMessage("La classe est requis."),
-    check("race").trim().notEmpty().withMessage("La race est requis."),
-    check("alignement")
-      .trim()
-      .notEmpty()
-      .withMessage("L'alignement' est requis."),
-    check("Traits.PersonalityTraits")
-      .trim()
-      .notEmpty()
-      .withMessage("Les Traits de personalités sont requis."),
-    check("Traits.Ideals")
-      .trim()
-      .notEmpty()
-      .withMessage("Les Ideaux sont requis."),
-    check("Traits.Bonds")
-      .trim()
-      .notEmpty()
-      .withMessage("Les Obligations sont requis."),
-    check("Traits.Flaws")
-      .trim()
-      .notEmpty()
-      .withMessage("Les Défauts sont requis."),
-    check("statistics.strength")
-      .trim()
-      .notEmpty()
-      .withMessage("La force est requise."),
-    check("statistics.dexterity")
-      .trim()
-      .notEmpty()
-      .withMessage("La dextérité est requise."),
-    check("statistics.constitution")
-      .trim()
-      .notEmpty()
-      .withMessage("La constitution est requise."),
-    check("statistics.intelligence")
-      .trim()
-      .notEmpty()
-      .withMessage("L'intelligence est requise."),
-    check("statistics.wisdom")
-      .trim()
-      .notEmpty()
-      .withMessage("La sagesse est requise."),
-    check("statistics.charisma")
-      .trim()
-      .notEmpty()
-      .withMessage("Le charisme est requis."),
-  ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-    try {
-      const body = req.body;
-      //   console.log(body);
-      await db.collection("characters").add(body);
-      const response = {
-        msg: "le personnage à bien été ajouter à la bibliothèque.",
-      };
-      res.status(201).json(response);
-    } catch (error) {
-      console.error("Erreur lors de la création du personnage :", error);
-      return res
-        .status(500)
-        .json({ error: "Erreur lors de la création du personnage." });
-    }
+
+router.post("/", validations, async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
   }
-);
+  try {
+    const body = req.body;
+    await db.collection("characters").add(body);
+    const response = {
+      msg: "le personnage à bien été ajouter à la bibliothèque.",
+    };
+    res.status(201).json(response);
+  } catch (error) {
+    console.error("Erreur lors de la création du personnage :", error);
+    return res
+      .status(500)
+      .json({ error: "Erreur lors de la création du personnage." });
+  }
+});
 
 /**
  * Route pour la page pour initialiser la bdd.
@@ -262,7 +277,11 @@ router.post("/dbinit", async (req, res) => {
 /**
  * route pour modifier un personnage
  */
-router.put("/:id", async (req, res) => {
+router.put("/:id", validations, async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() });
+  }
   try {
     const { id } = req.params;
     const { body } = req;
@@ -284,8 +303,12 @@ router.put("/:id", async (req, res) => {
 /**
  * route pour suprimer un personnage
  */
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", validId, async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() });
+    }
     const { id } = req.params;
     await db.collection("characters").doc(id).delete();
     const response = {
